Add note about primary domain delivery
When converting a single domain to multi-domain setup the primary domain seems to not be able to accept mail anymore. The common problem is that the original setup-kolab command creates multiple aliases which for whatever reason cause the hosted_triplet/hosted_duplet files to not find the domain. This adds a paragraph description the potential need for this change.